# frameworks/java/micronaut/Client/openapi.yml
openapi: 3.0.1
info:
  title: implementation
  version: 1.0.0
paths:
...
  /body/bind/medium:
    post:
      operationId: bindMedium
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        "200":
          description: bindMedium 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyRoutes.Bound_OrderRequest_"
  /body/bind/small:
    post:
      operationId: bindSmall
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        "200":
          description: bindSmall 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyRoutes.Bound_OrderRequest_"
  /body/validate/first-error:
    post:
      operationId: validateFirstError
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirstErrorOrder"
        required: true
      responses:
        "200":
          description: validateFirstError 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyRoutes.Bound_FirstErrorOrder_"
  /body/validate/medium:
    post:
      operationId: validateMedium
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        "200":
          description: validateMedium 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyRoutes.Bound_OrderRequest_"
  /body/validate/small:
    post:
      operationId: validateSmall
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderRequest"
        required: true
      responses:
        "200":
          description: validateSmall 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BodyRoutes.Bound_OrderRequest_"
...
  /cache/vary/many:
    get:
      operationId: varyMany
      parameters:
      - name: x-rb-channel
        in: header
        schema:
          type: string
          nullable: true
      - name: x-rb-region
        in: header
        schema:
          type: string
          nullable: true
      - name: x-rb-tenant
        in: header
        schema:
          type: string
          nullable: true
      responses:
        "200":
          description: varyMany 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payload"
  /cache/vary/one:
    get:
      operationId: varyOne
      parameters:
      - name: x-rb-tenant
        in: header
        schema:
          type: string
          nullable: true
      responses:
        "200":
          description: varyOne 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payload"
...
  /parameters/{one}/segment/literal:
    get:
      operationId: one
      parameters:
      - name: one
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: one 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Echoed_ParametersRoutes.One_"
  /parameters/{one}/with-second/{two}:
    get:
      operationId: two
      parameters:
      - name: one
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: two
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: two 200 response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Echoed_ParametersRoutes.Two_"
...
components:
  schemas:
    BodyRoutes.Bound_FirstErrorOrder_:
      required:
      - bytes
      - echo
      - fields
      type: object
      properties:
        fields:
          type: integer
          format: int32
        bytes:
          type: integer
          format: int64
        echo:
          type: object
      description: "What a bind or validate row answers: the order back, with the\
        \ leaves the handler found in it and the bytes it received."
    BodyRoutes.Bound_OrderRequest_:
      required:
      - bytes
      - echo
      - fields
      type: object
      properties:
        fields:
          type: integer
          format: int32
        bytes:
          type: integer
          format: int64
        echo:
          type: object
      description: "What a bind or validate row answers: the order back, with the\
        \ leaves the handler found in it and the bytes it received."
...
    Echoed_ParametersRoutes.One_:
      required:
      - count
      - echo
      - items
      - size
      type: object
      properties:
        size:
          type: string
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        echo:
          $ref: "#/components/schemas/ParametersRoutes.One"
      description: A payload with the values a handler bound written back beside its
        own fields.
    Echoed_ParametersRoutes.Two_:
      required:
      - count
      - echo
      - items
      - size
      type: object
      properties:
        size:
          type: string
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        echo:
          $ref: "#/components/schemas/ParametersRoutes.Two"
      description: A payload with the values a handler bound written back beside its
        own fields.
...
    FirstErrorOrder:
      required:
      - customerId
      - lines
      - status
      type: object
      properties:
        customerId:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
        status:
          minLength: 1
          type: string
        lines:
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/OrderRequest.Line"
      description: "The same body on the route that stops at the first bad field.\
        \ micronaut-validation reports every constraint a bean breaks and has no mode\
        \ that stops at the first. The group sequence replaces this record's default\
        \ group with one group per field, validated in order, and the validator stops\
        \ at the first group with a violation."
...
    Item:
      required:
      - category
      - id
      - inStock
      - name
      - priceCents
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        category:
          type: string
        priceCents:
          type: integer
          format: int32
        inStock:
          type: boolean
      description: "One row of items.large, and of every payload made from it."
...
    OrderRequest:
      required:
      - customerId
      - lines
      - status
      type: object
      properties:
        customerId:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
        status:
          minLength: 1
          type: string
        lines:
          minItems: 1
          type: array
          items:
            $ref: "#/components/schemas/OrderRequest.Line"
      description: "The body the bind and validate rows send, with the rules orderRequest\
        \ states as Bean Validation constraints. micronaut-validation's processor\
        \ compiles them into the record's introspection, and they run only where a\
        \ handler marks the parameter @Valid, so the bind routes bind the same record\
        \ and check nothing."
    OrderRequest.Line:
      required:
      - productId
      - qty
      type: object
      properties:
        productId:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
        qty:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
    ParametersRoutes.One:
      required:
      - one
      type: object
      properties:
        one:
          type: integer
          format: int32
    ParametersRoutes.Two:
      required:
      - one
      - two
      type: object
      properties:
        one:
          type: integer
          format: int32
        two:
          type: integer
          format: int32
    Payload:
      required:
      - count
      - items
      - size
      type: object
      properties:
        size:
          type: string
        count:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
      description: "items.small, items.medium or items.large."
...